Sync all processes in index file
